backend/requirements.txt
mysql.connector
flask
flask_cors